changed directory of  ErrorHandlerMiddleware, naming of exceptions and refactored ExampleErrorService into working example (more like prototype and without unit test)
